upto profile

+++ app.py
@@ -5,10 +5,13 @@ from datetime import datetime
 from werkzeug.security import generate_password_hash,check_password_hash
 from wtforms.widgets import TextArea
 from flask_login import UserMixin, login_user, LoginManager, login_required, logout_user, current_user
-from webforms import loginForm, PostForm, UserForm, PasswordForm, NameForm
+from webforms import loginForm, PostForm, UserForm, PasswordForm, NameForm, searchForm
+from flask_ckeditor import CKEditor
 
 #create a flask instance 
 app = Flask(__name__)
+#add ckeditor 
+ckeditor = CKEditor(app)
 #add database
 #sqlite db
 #app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///users.db'
@@ -46,24 +49,31 @@ def posts():
                          posts = posts)
 
 @app.route('/posts/delete/<int:id>')
+@login_required
 def delete_post(id):
   #get the post to delete from database
   post_to_delete = Posts.query.get_or_404(id)
-  try:
-    #delete the post from databae
-    db.session.delete(post_to_delete)
-    db.session.commit()
-    flash("Blog post was deleted!!")
-    #grab all the posts from the database
-    posts = Posts.query.order_by(Posts.date_posted)
-    return render_template("posts.html", posts = posts)
+  id = current_user.id
+  if id == post_to_delete.poster.id:
+    try:
+      #delete the post from databae
+      db.session.delete(post_to_delete)
+      db.session.commit()
+      flash("Blog post was deleted!!")
+      #grab all the posts from the database
+      posts = Posts.query.order_by(Posts.date_posted)
+      return render_template("posts.html", posts = posts)
 
-  except:
-    flash("Whoops! There was a problem while deleting the post, Try again...")
+    except:
+      flash("Whoops! There was a problem while deleting the post, Try again...")
 
-    posts = Posts.query.order_by(Posts.date_posted)
-    return render_template("posts.html", posts = posts)
-
+      posts = Posts.query.order_by(Posts.date_posted)
+      return render_template("posts.html", posts = posts)
+  else:
+      flash("You aren't authorized to delete that post.")
+      #grab all the posts from the database
+      posts = Posts.query.order_by(Posts.date_posted)
+      return render_template("posts.html", posts = posts)
 
 @app.route('/posts/edit/<int:id>', methods= ['POST', 'GET'])
 @login_required
@@ -73,7 +83,7 @@ def edit_post(id):
   if form.validate_on_submit():
     #get data to update from form
     post.title = form.title.data
-    post.author = form.author.data
+    # post.author = form.author.data
     post.slug = form.slug.data
     post.content = form.content.data
     #update database
@@ -81,13 +91,18 @@ def edit_post(id):
     db.session.commit()
     flash("Post has been updated!")
     return redirect(url_for('post',id = post.id))
-  #fill the form from data fromd database
-  form.title.data = post.title
-  form.author.data = post.author
-  form.slug.data = post.slug
-  form.content.data = post.content
-  return render_template('edit_post.html', form = form)
-
+  if current_user.id == post.poster_id:
+    #fill the form from data fromd database
+    form.title.data = post.title
+    # form.author.data = post.author
+    form.slug.data = post.slug
+    form.content.data = post.content
+    return render_template('edit_post.html', form = form)
+  else:
+    flash("You aren't authorized to edit that post.")
+    #grab all the posts from the database
+    posts = Posts.query.order_by(Posts.date_posted)
+    return render_template("posts.html", posts = posts)
   
 #create a post page
 @app.route('/add-post', methods=['GET', 'POST'])
@@ -96,15 +111,16 @@ def add_post():
   form = PostForm()
 
   if form.validate_on_submit():
+    poster = current_user.id
     post = Posts(title=form.title.data,
                  content = form.content.data,
-                 author = form.author.data,
+                 poster_id = poster,
                  slug = form.slug.data)
     #clear form 
     form.title.data = ''
     form.content.data = ''
     form.title.data = ''
-    form.author.data = ''
+    # form.author.data = ''
     form.slug.data = ''
     #add post to the database
     db.session.add(post)
@@ -142,6 +158,7 @@ def delete(id):
 
 #update database record
 @app.route('/update/<int:id>', methods=['GET','POST'])
+@login_required
 def update(id):
   form = UserForm()
   name_to_update = Users.query.get_or_404(id)
@@ -199,7 +216,37 @@ def add_user():
                          name = name,
                          our_users = our_users)
 
-  
+#create search funcion
+@app.route('/search', methods=['POST'])
+def search():
+  form = searchForm()
+  posts = Posts.query
+  if form.validate_on_submit():
+   #get data from submitted data
+   searched = form.searched.data
+   #Query the database
+   #posts = posts.filter_by(content = Posts.content.like('%' + searched + '%'))
+   posts = posts.filter_by(content = Posts.content.notlike('%' + searched + '%'))
+   posts = posts.order_by(Posts.title).all()
+   return render_template("search.html", form=form, searched = searched, posts = posts) 
+
+#pass stuff to navbar
+@app.context_processor
+def base():
+  form = searchForm()
+  return dict(form = form)
+
+#create admin page
+@app.route('/admin')
+@login_required
+def admin():
+  id = current_user.id 
+  if id == 17:
+    return render_template("admin.html")
+  else:
+    flash('Sorry, you must be admin to access the admin page.')
+    return redirect(url_for('dashboard'))
+
 #create login page
 @app.route('/login', methods = ['POST','GET'])
 def login():
@@ -239,6 +286,7 @@ def dashboard():
     name_to_update.username = request.form['username']
     name_to_update.email = request.form['email']
     name_to_update.favorite_color = request.form['favorite_color']
+    name_to_update.about_author = request.form['about_author']
     
     try:
       db.session.commit()
@@ -330,9 +378,11 @@ class Posts(db.Model):
   id = db.Column(db.Integer, primary_key = True)
   title = db.Column(db.String(255))
   content = db.Column(db.Text)
-  author = db.Column(db.String(255))
+  #author = db.Column(db.String(255))
   date_posted = db.Column(db.DateTime, default = datetime.utcnow)
   slug = db.Column(db.String(255))
+  #foreign key to link Users
+  poster_id = db.Column(db.Integer, db.ForeignKey('users.id'))
 
 #create user model
 class Users(db.Model, UserMixin):
@@ -341,9 +391,14 @@ class Users(db.Model, UserMixin):
   name = db.Column(db.String(200), nullable = False)
   email = db.Column(db.String(120), nullable = False, unique = True)
   favorite_color = db.Column(db.String(200))
+  about_author = db.Column(db.Text(500), nullable = True)
   date_added = db.Column(db.DateTime, default = datetime.utcnow)
+  # profile_pic = db.Column(db.String(), nullable = True)
   #do some password stuff!
   password_hash = db.Column(db.String(128))
+  #uers can have many Posts
+  posts = db.relationship('Posts', backref = 'poster')
+
   @property
   def password(self):
     raise AttributeError("Password is not a readable attribute!")
